# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nettplus Competitor Sentinel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Region {
  id: string;
  name: string;
  neighborhoods: string[];
}

export type CompetitorOfferLocalRelevance =
  (typeof CompetitorOfferLocalRelevance)[keyof typeof CompetitorOfferLocalRelevance];

export const CompetitorOfferLocalRelevance = {
  ALTA: "ALTA",
  MEDIA: "MEDIA",
  BAJA: "BAJA",
} as const;

export interface CompetitorOffer {
  id: string;
  detectedDate: string;
  competitor: string;
  adImageUrl: string;
  adCaption: string;
  plan: string;
  price: number;
  speedMbps: number;
  inferredLocation: string;
  specificLocation: string;
  offerType: string;
  sourceUrl: string;
  localRelevance: CompetitorOfferLocalRelevance;
  isNew: boolean;
}

export interface PriceComparison {
  competitor: string;
  avgPricePerMbps: number;
  avgPrice: number;
  offerCount: number;
  minPrice: number;
  maxPrice: number;
}

export interface MarketShare {
  competitor: string;
  offerCount: number;
  percentage: number;
  color: string;
}

export interface SwotAnalysis {
  strengths: string[];
  weaknesses: string[];
  opportunities: string[];
  threats: string[];
  lastUpdated: string;
}

export type MapPointIntensity =
  (typeof MapPointIntensity)[keyof typeof MapPointIntensity];

export const MapPointIntensity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface MapPoint {
  lat: number;
  lng: number;
  competitor: string;
  plan: string;
  price: number;
  neighborhood: string;
  intensity: MapPointIntensity;
}

export interface DashboardSummary {
  totalOffersDetected: number;
  highRelevanceOffers: number;
  competitorsTracked: number;
  avgMarketPrice: number;
  lastRefresh: string;
  alertsCount: number;
  newOffersCount: number;
  selectedRegionName: string;
}

export type GetCompetitorOffersParams = {
  competitor?: string;
  location?: string;
  neighborhoods?: string;
  dateFrom?: string;
  dateTo?: string;
};

export type GetPriceComparisonParams = {
  location?: string;
  neighborhoods?: string;
};

export type GetMarketShareParams = {
  location?: string;
  neighborhoods?: string;
};

export type GetSwotAnalysisParams = {
  location?: string;
  competitor?: string;
};

export type GetMapPointsParams = {
  location?: string;
};

export type GetCompetitorSummaryParams = {
  location?: string;
  neighborhoods?: string;
};