# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nettplus Competitor Sentinel API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: competitors
    description: Competitor intelligence data
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /competitors/regions:
    get:
      operationId: getRegions
      tags: [competitors]
      summary: Get regions and neighborhoods
      description: Returns the regional structure of Quito with sub-neighborhoods
      responses:
        "200":
          description: Regions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"

  /competitors/offers:
    get:
      operationId: getCompetitorOffers
      tags: [competitors]
      summary: Get competitor offers
      description: Returns simulated competitor ad/offer data with geographic relevance scoring
      parameters:
        - name: competitor
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: neighborhoods
          in: query
          required: false
          schema:
            type: string
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of competitor offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompetitorOffer"

  /competitors/price-comparison:
    get:
      operationId: getPriceComparison
      tags: [competitors]
      summary: Get price comparison data
      description: Returns price per mega comparison across competitors
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: neighborhoods
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Price comparison data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceComparison"

  /competitors/market-share:
    get:
      operationId: getMarketShare
      tags: [competitors]
      summary: Get market share data
      description: Returns offer count per competitor for the selected region
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: neighborhoods
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Market share data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketShare"

  /competitors/swot:
    get:
      operationId: getSwotAnalysis
      tags: [competitors]
      summary: Get SWOT analysis
      description: Returns dynamic SWOT analysis based on filters
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: competitor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: SWOT analysis data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwotAnalysis"

  /competitors/map-points:
    get:
      operationId: getMapPoints
      tags: [competitors]
      summary: Get map data points
      description: Returns geographic data points for competitor offers
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Map data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapPoint"

  /competitors/summary:
    get:
      operationId: getCompetitorSummary
      tags: [competitors]
      summary: Get summary stats
      description: Returns high-level summary statistics for the dashboard
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: neighborhoods
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Region:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        neighborhoods:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - neighborhoods

    CompetitorOffer:
      type: object
      properties:
        id:
          type: string
        detectedDate:
          type: string
          format: date
        competitor:
          type: string
        adImageUrl:
          type: string
        adCaption:
          type: string
        plan:
          type: string
        price:
          type: number
        speedMbps:
          type: number
        inferredLocation:
          type: string
        specificLocation:
          type: string
        offerType:
          type: string
        sourceUrl:
          type: string
        localRelevance:
          type: string
          enum: [ALTA, MEDIA, BAJA]
        isNew:
          type: boolean
      required:
        - id
        - detectedDate
        - competitor
        - adImageUrl
        - adCaption
        - plan
        - price
        - speedMbps
        - inferredLocation
        - specificLocation
        - offerType
        - sourceUrl
        - localRelevance
        - isNew

    PriceComparison:
      type: object
      properties:
        competitor:
          type: string
        avgPricePerMbps:
          type: number
        avgPrice:
          type: number
        offerCount:
          type: number
        minPrice:
          type: number
        maxPrice:
          type: number
      required:
        - competitor
        - avgPricePerMbps
        - avgPrice
        - offerCount
        - minPrice
        - maxPrice

    MarketShare:
      type: object
      properties:
        competitor:
          type: string
        offerCount:
          type: number
        percentage:
          type: number
        color:
          type: string
      required:
        - competitor
        - offerCount
        - percentage
        - color

    SwotAnalysis:
      type: object
      properties:
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        opportunities:
          type: array
          items:
            type: string
        threats:
          type: array
          items:
            type: string
        lastUpdated:
          type: string
      required:
        - strengths
        - weaknesses
        - opportunities
        - threats
        - lastUpdated

    MapPoint:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        competitor:
          type: string
        plan:
          type: string
        price:
          type: number
        neighborhood:
          type: string
        intensity:
          type: string
          enum: [high, medium, low]
      required:
        - lat
        - lng
        - competitor
        - plan
        - price
        - neighborhood
        - intensity

    DashboardSummary:
      type: object
      properties:
        totalOffersDetected:
          type: number
        highRelevanceOffers:
          type: number
        competitorsTracked:
          type: number
        avgMarketPrice:
          type: number
        lastRefresh:
          type: string
        alertsCount:
          type: number
        newOffersCount:
          type: number
        selectedRegionName:
          type: string
      required:
        - totalOffersDetected
        - highRelevanceOffers
        - competitorsTracked
        - avgMarketPrice
        - lastRefresh
        - alertsCount
        - newOffersCount
        - selectedRegionName