# lib/api-zod/src/generated/types/competitorOffer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nettplus Competitor Sentinel API
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitorOfferLocalRelevance } from "./competitorOfferLocalRelevance";

export interface CompetitorOffer {
  id: string;
  detectedDate: Date;
  competitor: string;
  adImageUrl: string;
  adCaption: string;
  plan: string;
  price: number;
  speedMbps: number;
  inferredLocation: string;
  specificLocation: string;
  offerType: string;
  sourceUrl: string;
  localRelevance: CompetitorOfferLocalRelevance;
  isNew: boolean;
}