# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nettplus Competitor Sentinel API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalOffersDetected: number;
  highRelevanceOffers: number;
  competitorsTracked: number;
  avgMarketPrice: number;
  lastRefresh: string;
  alertsCount: number;
  newOffersCount: number;
  selectedRegionName: string;
}