# lib/api-zod/src/generated/types/region.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nettplus Competitor Sentinel API
 * OpenAPI spec version: 0.1.0
 */

export interface Region {
  id: string;
  name: string;
  neighborhoods: string[];
}